Move RNG functions ext/stanard to ext/random
