+++ ext/random/config.w32
@@ -1,0 +1,2 @@
+EXTENSION('random', 'random.c', false /* never shared */, '/DZEND_ENABLE_STATIC_TSRMLS_CACHE=1');
+PHP_INSTALL_HEADERS("", "ext/random");
