+++ ext/random/tests/bug46587.phpt
