+++ ext/random/tests/bug75170.phpt
