+++ ext/random/tests/bug75514.phpt
