+++ ext/random/tests/mt_rand_basic.phpt
