+++ ext/random/tests/mt_rand_value.phpt
