+++ ext/random/tests/rand.phpt
