+++ ext/random/tests/rand_basic.phpt
