+++ ext/random/tests/rand_inverted_order.phpt
