+++ ext/random/tests/random_bytes.phpt
