+++ ext/random/tests/random_int_error.phpt
