+++ ext/session/session.c
@@ -37,10 +37,10 @@
 #include "php_variables.h"
 #include "php_session.h"
 #include "session_arginfo.h"
-#include "ext/standard/php_random.h"
+#include "ext/random/php_random.h"
 #include "ext/standard/php_var.h"
 #include "ext/date/php_date.h"
-#include "ext/standard/php_lcg.h"
+#include "ext/random/php_random.h"
 #include "ext/standard/url_scanner_ex.h"
 #include "ext/standard/info.h"
 #include "zend_smart_str.h"
