--- ext/standard/php_mt_rand.h
@@ -1,40 +1,0 @@
-/*
-   +----------------------------------------------------------------------+
-   | Copyright (c) The PHP Group                                          |
-   +----------------------------------------------------------------------+
-   | This source file is subject to version 3.01 of the PHP license,      |
-   | that is bundled with this package in the file LICENSE, and is        |
-   | available through the world-wide-web at the following url:           |
-   | https://www.php.net/license/3_01.txt                                 |
-   | If you did not receive a copy of the PHP license and are unable to   |
-   | obtain it through the world-wide-web, please send a note to          |
-   | license@php.net so we can mail you a copy immediately.               |
-   +----------------------------------------------------------------------+
-   | Authors: Rasmus Lerdorf <rasmus@php.net>                             |
-   |          Zeev Suraski <zeev@php.net>                                 |
-   |          Pedro Melo <melo@ip.pt>                                     |
-   |          Sterling Hughes <sterling@php.net>                          |
-   |                                                                      |
-   | Based on code from: Shawn Cokus <Cokus@math.washington.edu>          |
-   +----------------------------------------------------------------------+
- */
-
-#ifndef PHP_MT_RAND_H
-#define PHP_MT_RAND_H
-
-#include "php_lcg.h"
-#include "php_rand.h"
-
-#define PHP_MT_RAND_MAX ((zend_long) (0x7FFFFFFF)) /* (1<<31) - 1 */
-
-#define MT_RAND_MT19937 0
-#define MT_RAND_PHP 1
-
-PHPAPI void php_mt_srand(uint32_t seed);
-PHPAPI uint32_t php_mt_rand(void);
-PHPAPI zend_long php_mt_rand_range(zend_long min, zend_long max);
-PHPAPI zend_long php_mt_rand_common(zend_long min, zend_long max);
-
-PHP_MINIT_FUNCTION(mt_rand);
-
-#endif	/* PHP_MT_RAND_H */
